# swagger/openapi.yaml
openapi: 3.0.0
info:
  title: Geek Marketplace API
  version: 1.0.0

paths:
  /api/post/{id}/service:
    post:
      summary: Activate VIP or TOP service
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                type:
                  type: string
                  enum: [vip, top]
      responses:
        '200':
          description: Service activated
        '400':
          description: Invalid type
        '401':
          description: Unauthorized
        '404':
          description: Advert not found

    put:
      summary: Extend existing service
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                type:
                  type: string
                  enum: [vip, top]
      responses:
        '200':
          description: Service extended
        '400':
          description: Invalid type
        '401':
          description: Unauthorized
        '404':
          description: Service not found

    get:
      summary: Get service info for advert
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Service info returned
        '401':
          description: Unauthorized
        '404':
          description: Service not found

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT